backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
redis[asyncio]==5.2.1
python-jose[cryptography]==3.3.0
boto3==1.35.87
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.14
httpx==0.28.1
razorpay==1.4.2
sentry-sdk[fastapi]==2.19.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
ruff==0.8.4
